--- a/package.json
+++ b/package.json
@@ -35,6 +35,7 @@
     "@babel/preset-react": "7.18.6",
     "@edx/eslint-config": "3.1.1",
     "@edx/new-relic-source-map-webpack-plugin": "1.0.2",
+    "@fullhuman/postcss-purgecss": "^5.0.0",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.10",
     "@svgr/webpack": "6.5.1",
     "autoprefixer": "10.4.13",