This is ready again for review, see below for testing detials and result

~~[Update] 18.01.22 As of testing with olive MFE learning, the CSS output was missing a lot of stuff. Converted to draft again, need to double check the correct settings~~


 The added plugin, reduce the CSS file by removing unused CSS
 The plugin is kinda recommended by bootstrap[^1], reducing
 unused CSS is generally recommended as well enchance web
 perforamce[^2]

 [^1]: https://getbootstrap.com/docs/5.2/customize/optimize/#unused-css
 [^2]: https://web.dev/css-web-vitals/#critical-css

 Related issue: openedx/frontend-wg/issues/138

This is still work on progress, I need to do further testing, to ensure this change doesn't break styling or any other thing