@ghassanmas 2023-01-18
[Update]: 

Testing: I have tested this version on frontend-app-learning, by installing this version of frontend-build: see this PR https://github.com/openedx/frontend-app-learning/pull/1047 

The effect of this is that the css way more than expected. I don't rememeber how I test it when I opened the draft, but in any case this cannon be merged as it is.  I am going convert it to draft again, and when time allow I will dig into the correct settings. s