@ghassanmas 2023-01-18
[Update] I think I made it to work with learning MFE olive, CSS file size was decreased 70% (few hundred kilo bytes). I had to completely set it in a different way, please check the expected diff at https://github.com/ghassanmas/frontend-build/pull/1 . If that can of change acceptable I would tidy up and update this PR accordingly,