@ghassanmas 2023-01-18
# The test on MFE Learning

## Before This change 

**CSS File size**: 132KB GIZ /883K/Original

<img width="760" alt="image" src="https://user-images.githubusercontent.com/16361375/213305397-6f9019b0-16c9-4c8a-8537-d72a5aba4301.png">

Live link: https://apps.olive.zaat.dev/learning/course/course-v1:zaatdev+101+1/home


## After This change 

<img width="760" alt="image" src="https://user-images.githubusercontent.com/16361375/213304168-2b5cedc2-8674-49d0-a9d2-27d570ae19d1.png">

**CSS File size**:  27KB GIZPPED/  278K Orginal 

Live Link: https://learning.olive.zaat.dev/course/course-v1:zaatdev+101+1/home


## Concolusion 

 Increaing [lighthouse](https://web.dev/measure/) performance by 20%