@abdullahwaheed 2023-01-19
Its improving the build size but build time is affecting greatly with these changes. I have tested it on multiple MFEs and following are the results:
**frontend-app-account**:
    without css optimization -> build time: 65974 ms
    with optimization -> build time: 150820 ms

**frontend-app-learner-portal-enterprise**:
    without css optimization -> build time: 73184 ms
    with optimization -> build time: 268651 ms

**frontend-app-learning**:
    without css optimization -> build time: 121364 ms
    with optimization -> build time: 239722 ms


@arbrandes what are your thoughts on this?