@ghassanmas 2023-01-19
Yes I expected that would occuar, but for me I think the priority comes for end user which is leaner. 

For example at 2U/edx you do one build for milllions of users, compare those few seconds with a fraction of second of for million of users.