@arbrandes 2023-01-24
I have to be honest: I don't much like CSS removers.  There's too much risk for too little reward.  This article outlines some of the dangers:

https://css-tricks.com/how-do-you-remove-unused-css-from-a-site/#aa-part-of-a-build-process

The added build time is also bad.  We're trying to reduce it for many reasons, and this will make it much worse.  I don't think saving 100KB of downloads is worth it.

What I'd be more interested in is ways to avoid having to insert framework CSS when it's not necessary in the first place, possibly by picking more modular frameworks.