@ghassanmas 2023-01-31
@itsjeyd we have discussed this in the last FWG, https://openedx.atlassian.net/wiki/spaces/COMM/pages/3645964289/2023-01-26+Frontend+Working+Group+Meeting+Minutes. And I think @arbrandes want to check with folks opinion before merging it.