@itsjeyd 2023-02-08
Thanks for the info @ghassanmas!

As of today, what do you think will be the next steps for this PR? Is it still waiting for results from this:

> I think @arbrandes want to check with folks opinion before merging it.

... Or are you considering to follow @regisb's suggestion and reimplement the changes as a plugin (in which case it might make sense to close this PR)?