@ghassanmas 2023-02-08
Thinking about this, I came with two scnario: 
1. implement a featuer flag to enable/disable adding this plugin 
   - it can be by default off which should not affect tutor/tutor-mfe
1.  Doing a tutor plugin for that is indeed possiblem but it would make it a bit of hassle to maintain ** a plugin and a fork of front-build**

Overall the bigger picture of this I think we ideally tutor-mfe can provide an image which is built once for all other providers, and it could have feature flag of which when its turned on **it would make build faster but not optimize the build size** [^1]. The scinario I have in mind is that developer would keep rebuilding the image until they see the changes they want, and then once it looks good we can advise them to rebuild while turnning all optimization flag

 [^1]: For that to we might not only skip this plugin but also other plugin that make the build slow, see overhangio/tutor-mfe/issues/88 for example.