@regisb 2023-02-10
> Doing a tutor plugin for that is indeed possible but it would make it a bit of hassle to maintain a plugin and a fork of front-build

Could we avoid forking frontend-build? If I understand correctly all we have to do is to modify webpack.config.js, right? So we could ship our own version of that file, just like we do in development.

That being said, my initial knee-jerk reaction is the same as Adolfo: CSS removal looks tricky and I'd be afraid to break too many things.