@regisb 2023-02-10
I realize now that my [previous comment](https://github.com/openedx/frontend-build/pull/266#issuecomment-1416056973) was stupid. I initially thought that this issue was part of the tutor-mfe plugin, and not frontend-build :sweat: Thus my comment about making this change part of tutor-mfe does not really make sense anymore...

IMHO the compressed size decrease is pretty serious, and it's worth considering. But I'm wondering if the proposed fix is not the tree that hides the forest. I mean look at the numbers: are we seriously generating a CSS file that weighs 883kB? What for? Clearly, this pull request shows that there is a ton of crap in this file (69% to be precise). The word "bloated" comes to mind. Why is there so much stuff in this file in the first place? I'm afraid that if we use a plugin to purge this file we'll just blindingly adding more and more stuff, and in a couple months/years down the road we'll be facing the same problem again.