@jmbowman 2023-02-10
Building on the last comment - can we use the purgecss output to identify the stuff being ripped out, and try to avoid including it in the first place?  Maybe via appropriate Bootstrap or Paragon configuration options, etc.